requirements/requirements_bot.txt
aiogram==2.24
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
colorama==0.4.6
dnspython==2.3.0
frozenlist==1.3.3
fuzzywuzzy==0.18.0
h11==0.14.0
Levenshtein==0.20.9
mongoengine==0.26.0
numpy==1.24.2
opencv-python==4.7.0.72
Pillow==9.4.0
pychalk==2.0.1
pydantic==1.10.5
pymongo==4.3.3
pytesseract==0.3.10
python-Levenshtein==0.20.9
pytz==2022.7.1
PyYAML==6.0
rapidfuzz==2.13.7
requests==2.28.2
schedule==1.1.0